refactor(jobs): ship PostgreSQL only, drop the SQL Server implementation
SQL Server support returns with the wider cross-module provider work; until
then a second provider costs a duplicate migration, snapshot, context,
factory and entity configuration that nothing exercises. Removing it also
takes the duplication that the quality gate was flagging, since the two
generated migration bodies were near-identical by construction.

Nothing regresses today: no feature enqueues jobs yet, so a SQL Server
deployment loses nothing. Rather than fail startup there — which would take
down installations that never touch the queue — the module registers a
stand-in queue that throws with a message naming the constraint, so a future
consumer that forgets it gets an explanation instead of a dependency
resolution error.

The provider-split context stays. EF Core binds migrations to the context
type, so collapsing JobsPostgreSqlDbContext into a single JobsDbContext now
would mean rebuilding the migration history when a second provider arrives;
keeping the split makes that addition purely additive. The entity
configuration is collapsed instead, because splitting it again later is free.
